0.4.2.0

+++ Misc/New-DesktopShortcut.ps1
@@ -7,13 +7,16 @@
         [string]$WorkingDirectory = "", 
         [string]$IconLocation = "", 
         [string]$Arguments = "",
-        [ValidateSet('None','Desktop','StartMenu','Startup','CommonStartMenu')]
+        [ValidateSet('None','Desktop','StartMenu','Startup','CommonDesktop','CommonStartMenu','CommonStartup')]
         [string]$shortcuts = "Desktop",
         [switch]$RunAsAdministrator = $true
     )
     if ($shortcuts -ne "None") {
         
-        if ($shortcuts -eq "Desktop" -or $shortcuts -eq "Startup") {
+        if ($shortcuts -eq "Desktop" -or 
+            $shortcuts -eq "CommonDesktop" -or 
+            $shortcuts -eq "Startup" -or 
+            $shortcuts -eq "CommonStartup") {
             $folder = [Environment]::GetFolderPath($shortcuts)
         } else {
             $folder = Join-Path ([Environment]::GetFolderPath($shortcuts)) "NavContainerHelper"
@@ -22,29 +25,31 @@
             }
         }
 
-        $filename = Join-Path $folder "$Name.lnk"
-        if (Test-Path -Path $filename) {
-            Remove-Item $filename -force
-        }
-    
-        $Shell =  New-object -comobject WScript.Shell
-        $Shortcut = $Shell.CreateShortcut($filename)
-        $Shortcut.TargetPath = $TargetPath
-        if (!$WorkingDirectory) {
-            $WorkingDirectory = Split-Path $TargetPath
-        }
-        $Shortcut.WorkingDirectory = $WorkingDirectory
-        if ($Arguments) {
-            $Shortcut.Arguments = $Arguments
-        }
-        if ($IconLocation) {
-            $Shortcut.IconLocation = $IconLocation
-        }
-        $Shortcut.save()
-        if ($RunAsAdministrator) {
-            $bytes = [System.IO.File]::ReadAllBytes($filename)
-            $bytes[0x15] = $bytes[0x15] -bor 0x20 #set byte 21 (0x15) bit 6 (0x20) ON
-            [System.IO.File]::WriteAllBytes($filename, $bytes)
+        if ($folder) {
+            $filename = Join-Path $folder "$Name.lnk"
+            if (Test-Path -Path $filename) {
+                Remove-Item $filename -force
+            }
+        
+            $Shell =  New-object -comobject WScript.Shell
+            $Shortcut = $Shell.CreateShortcut($filename)
+            $Shortcut.TargetPath = $TargetPath
+            if (!$WorkingDirectory) {
+                $WorkingDirectory = Split-Path $TargetPath
+            }
+            $Shortcut.WorkingDirectory = $WorkingDirectory
+            if ($Arguments) {
+                $Shortcut.Arguments = $Arguments
+            }
+            if ($IconLocation) {
+                $Shortcut.IconLocation = $IconLocation
+            }
+            $Shortcut.save()
+            if ($RunAsAdministrator) {
+                $bytes = [System.IO.File]::ReadAllBytes($filename)
+                $bytes[0x15] = $bytes[0x15] -bor 0x20 #set byte 21 (0x15) bit 6 (0x20) ON
+                [System.IO.File]::WriteAllBytes($filename, $bytes)
+            }
         }
     }
 }
